requirements.txt
duckdb==1.1.3
fastapi==0.115.8
langchain_core==0.3.25
langchain_openai==0.2.12
# langchain_pinecone==0.2.0
langgraph==0.2.59
pandas==1.5.3
pydantic==2.10.3
python-dotenv==1.0.1
streamlit==1.40.0
uvicorn==0.34.0
tabulate==0.9.0
langchain_community==0.3.12
rapidfuzz==3.12.2
sqlparse==0.5.3
sqlalchemy==2.0.29
psycopg2-binary==2.9.10
python-jose==3.3.0
passlib==1.7.4